Fix type error

+++ cpp/src/groupby/hash/compute_aggregations.cuh
@@ -174,7 +174,7 @@ rmm::device_uvector<cudf::size_type> compute_aggregations(
     auto const stride = GROUPBY_BLOCK_SIZE * grid_size;
     thrust::for_each_n(rmm::exec_policy_nosync(stream),
                        thrust::make_counting_iterator(int64_t{0}),
-                       num_rows * flattened_values.num_columns(),
+                       num_rows * static_cast<int64_t>(flattened_values.num_columns()),
                        global_memory_fallback_fn{global_set_ref,
                                                  *d_values,
                                                  *d_sparse_table,
